# backend/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: AlgoForce CapitalOS API
  version: 0.1.0
paths:
...
  /ventures/{ventureId}:
    get:
      summary: Fetch the full Venture Object read model.
      parameters:
        - name: ventureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Venture Object.
  /ventures/{ventureId}/validate:
    post:
      summary: Validate idea, capital, and genome thresholds.
      responses:
        "200":
          description: VentureValidatedEvent emitted.
        "409":
          description: TransitionBlockedEvent emitted.
...